First publish
Code to list and loop throughout the files in a folder to extract and process the data. Then the validated data is being filled out into a file or passed to an API.
Code to list and loop throughout the files in a folder to extract and process the data. Then the validated data is being filled out into a file or passed to an API.

Code to list and loop throughout the files in a folder to extract and process the data. Then the validated data is being filled out into a file or passed to an API.